chore: Moved player to main isolate
There was no real reason to keep it in the background isolate.
This change allows for massively shortened callback code, easier
understanding of what's going on when an alarm goes off through log
messages in the main isolate and cleaner handling of things.

The reason this was in the background isolate originally was connected
to the naive belief that I could control the main isolate from the
background one. With the forced usage of a file as bridge to detect an
alarm, this logic can all be moved over to the main isolate now.

If nothing else did before, this has a good change to fix a lot of
things if not everything in #5